+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## v4.1.3
 
+- 新增Borax内置 `basic1`，与`basic` 相比，增加其他22个节气节日
 - 修正日历界面类 `CalendarFrame` 无法加载自定义节日源的bug [#62](https://github.com/kinegratii/borax/issues/62)
 - 新增 `FestivalLibrary` 独立文档页面